Always return a Version object for indirect gomodule dependencies
When dealing with an indirect gomodules dependency, we were retuning a
string value that came from the Dependency object. We use this version
to compare with a Dependency object in
common/lib/dependabot/update_checkers/base.rb:216, which resulted in an
ArgumentError. We already return a Version object when the dependency is
top_level.